refactor: split admin.service.ts analytics into 3 domain sub-services #342
